Feat : Implementation - Charactor ReverseAnimation Move and Jump
캐릭터 이미지 리소스 반전 작업 및 애니메이션 구현 (이동 점프)

+++ ANGVIK/MapColliderManager.cpp
@@ -47,7 +47,7 @@ bool MapColliderManager::autoMove(POINTFLOAT pos, RECT shape, float moveSpeed, P
 
 POINTFLOAT MapColliderManager::Move(POINTFLOAT pos, RECT shape, float moveSpeed, int dir, POINT bodySize)
 {
-	int height = 0;
+	float height = 0.0f;
 
 	for (int i = 1; i < bodySize.y; ++i)
 	{
